Implement the pattern store
The idea here is that I need a big global store of all patterns in the
system.  This is currently implemented as a bunch of static stuff
because I want it to be able to be accessed by library constructors so
I don't have to have a big list somewhere that contains every possible
pattern.

Additionally, I've added a skeleton for "chisel-patterns", which
generates a list of all the patterns that this code knows about.  I'm
currently trying to implement this by just writing a simple counter
pattern.