Use libflo for circuit graphs
I realized that I'd be duplicating a whole bunch of libflo in order to
deal with all the graph stuff so instead I just went ahead and used
it.  Note that this is going to require a whole bunch of changes to
the actual implementation of libflo, but I guess I'll have to live
with that for now... :).