Fix operation "operator+" to emit constants
I'm not exactly sure why the type system didn't catch this, but for
some reason I was constructing an operation with an integer instead of
a constant node object, which was causing libflo to silently not emit
the node.  I'm assuming this is some sort of implicit constructor
issue?

The fact that this can happen is almost certainly a bug in
libflo... :(