Add a test of "flo-patterns"
There are two sorts of tests here: one tests that the output of
"flo-patterns --list" has not changed, and another tests every example
pattern.  The first test is just designed to make sure that we have a
known list of all patterns for testing (which usually requires running
the executable).  The second test makes sure that each pattern works
all by itself.