Use "flo-llvm --torture"
This is a special mode of operation that generates VCD files that
coorespond to what Torture expects.  Essentially the idea is that
Chisel appears to generate a bunch of junk (internal registers, for
example) that we don't necessairially even know how to track with
torture.  This flag prevents those nodes from being emitted.