Add "one_sided_and_tree"
This is probably poorly named.  The general idea is that this just
builds a big linear AND of a single value, done many times.  Something
like

  (((in & in) & in) & in) ...

This has a few advantages: it's simple (the input is the output), it's
single cycle (all combinational logic), and it's not efficient on
DREAMER (so it serves as a test case for flo-opt).