Accept different "--seed" values for *-torture
This essentially allows different circuits to be specified on the
command-line.  Note that this currently defaults to a seed of 0, which
means that every time you run flo-torture you'll be getting exactly
the same circuit.

The way that seeding currently works means that if pretty much
anything changes about the list of circuits that can be generated then
a completely different mapping of seeds to outputs will result.