Generate Chisel code in addition to Flo code
In order for chisel-torture to be useful it needs to generate Chisel
code (which really means Scala code).  It currently doesn't do a whole
lot: it simply converts each Flo operation to a Chisel operation.

Essentially this is just a test for Chisel backends: it only emits a
single module and it only generats Bits (with a whole bunch of casts
to UInt to make arithmetic work).

I also run the same tests as run through the Flo backend.  Note that
there's a bit of a bug here in that Chisel appears to emit spurious
nodes so I can't vcddiff in both directions, but I guess that's OK for
now.