Remove some unused code from format_chisel
This is actually just a duplicate of the code I had in node, which
formats a node's name as how it would appear in a Chisel class.  I
decided it was better to put this in node (just like I do in
flo-llvm), but I forgot to get rid of this code.