Add a variable right shift test
This is almost exactly the same as the right shift test but it does a
variable right shift (as opposed to a constant right shift).  This
exposes a bit of undefined behavior I've known about for a while in
Chisel, but I've disabled that version of the test to avoid any
failures for now.