Add more LSH patterns
These aim to discover some bugs in the multi-word expander.  The
general idea is that LSH can be used to truncate and thus needs to
have all operations kept at a proper width throughout the whole
process in order to make sure that no bits are lost in the middle.